refactor: extract auth feature folder
Move auth-form widget, wordlist, and migrateAnonymousData into
src/app/features/auth/. Keep migrateAnonymousData re-exported
from @/schema for compatibility.
