+++ src/app/features/auth/index.ts
@@ -1,0 +1,3 @@
+export { AuthForm, AuthDialog } from "./widgets/auth-form"
+export { wordlist } from "./lib/wordlist"
+export { migrateAnonymousData } from "./lib/migrate-anonymous-data"
