+++ src/app/routes/invite.tsx
@@ -3,7 +3,7 @@ import { createFileRoute, useNavigate, Link } from "@tanstack/react-router"
 import { useAccount, useIsAuthenticated } from "jazz-tools/react"
 import { FileText, FolderOpen, AlertCircle, Loader2 } from "lucide-react"
 import { Button } from "@/components/ui/button"
-import { AuthDialog } from "@/components/auth-form"
+import { AuthDialog } from "@/app/features/auth"
 import { UserAccount } from "@/schema"
 import { acceptDocumentInvite, type DocInviteData } from "@/lib/documents"
 import { acceptSpaceInvite, type SpaceInviteData } from "@/lib/spaces"
